## Description
- Fixed issue where during filtering input when input key matched any of the keybinding it performed key binding action ( for example when entering "a" during filter it would toggle to create item screen). Updated keyHandlers to skip key handing when list is in filtering state.
- Added new keybinding "z" that toggles/hides completed items. Updated `ItemsList` model interface to let implementing models define `HideValue bool` that also applies to filtering when pressing keybinding.
- Fixed `ci` workflow to actually compile with `CGO_ENABLED=1` and building on actuall machines instead of having to deal with cross-compilation.

+++ internal/tui/itemsbytag/handlers.go
@@ -147,6 +147,9 @@ func (m *MenuModel) HandleInputState(msg tea.Msg) (textinput.Model, []tea.Cmd) {
 // HandleKeyInput processes key inputs not handles by list.Model
 func (m *MenuModel) HandleKeyInput(msg tea.KeyMsg) tea.Cmd {
 	var cmd tea.Cmd
+	if m.List.SettingFilter() {
+		return nil
+	}
 	switch {
 	case key.Matches(msg, m.Keys.DeleteItem):
 		cmd = m.DeleteItem()
